ctdTagApp() gives notification if 'i' won't work because out of focus
I need to find a way to either shift the focus, or -- I think better --
provide some GUI indication of whether the plot is in focus.

Or is this just an R-shiny "feature", that operations remove focus from
the plot panel?

Maybe I need to do some web searching, because this has bothered me in
various R-shiny apps for years.