improve cursor response BUT...
... I did a lot of web searching and don't see how to change the cursor
type when the mouse is in/out of the plot.  The code near "i" is where
to look for what I've tried.  As with so much to do with shiny, I don't
know how to figure out the CSS.  Sure, I can use an inspector in my web
browser, but can I override the cursor set there, and even if so, how do
I know whether they might change the class name for an element?