rename createDatabase() to useDatabase()
This name more accurately describes what the function does: create a
database if the named file does not exist, or check that the named file
is in the most up-to-date format, modifying it as necessary.  The latter
is done by calling `updateDatabase()`.

Another change is that, previously, all database-touching functions
started by calling `updateDatabase()`.  This is no longer done. The
intension is that users start by calling `useDatabase()`, and then move
on to other functions.  Removing this step of checking the database
version on every contact may double the speed of the work. Of course,
that will not matter for interactive work, such as that of
`ctdTagApp()`, but it might be helpful in automated work.  And since I'm
doing automated work at the moment, this change is evidence of eating
the dog food.